Create a new javac* task
It's based upon the built-in javac task, but extended to use the code
reloading compiler stuff from ztellman's virgil library.

This won't quite work though, because virgil does two things:

- compiles the compiled code in-memory only
- immediately loads the code with the DynamicClassLoader from
  clojure.lang

This is a clever way of making these classes dynamically reload and
accesible to Clojure code, but it doesn't help when trying to fully
reload Java classes, and it also makes it so this compilation process
isn't useful for actual compiles.

This second issue is fine, because this can be a development task only.
Use the normal javac for production builds. (Eventually, this should
probably be integrated with the built-in javac task much better, by
adding a flag to activate the purely dynamic reloading behavior).

The question now is how do I get these dynamically loaded and reloadable
classes to show up where cljunit can find them, and also get them to be
loaded with the correct ClassLoader so that they will be reloaded
properly?

I think the answer is (hopefully) that I need to pass the same class
loader used for this compilation into cljunit, which means it needs to
run not inside a pod, and then I need to force Reflections to use the
DynamicClassLoader.  With the classes only existing in memory in the
DCL, this should work just fine if I pass the DCL into reflections, when
it's not inside of a boot pod.